platform_system/requirements.txt
uwsgi
MySQL-python
django-widget-tweaks
